fix(react-langchain): keep the converter subpath declarations self-contained
