+++ .changeset/langchain-converter-seam.md
@@ -3,4 +3,4 @@
 "@assistant-ui/react-langgraph": patch
 ---
 
-refactor: share the langchain converter layer with react-langgraph. @langchain/react becomes an optional peer of react-langchain.
+refactor: share the langchain converter layer with react-langgraph via the new @assistant-ui/react-langchain/converter subpath. @langchain/react becomes an optional peer of react-langchain.
